Move most of nuke's actions into their own module
